implemented login component. Updated file architecture
