+++ assets/arrays.js
@@ -1,0 +1,3 @@
+export let cartArray = [];
+export let favBooks = [];
+
