+++ styles.css
@@ -466,6 +466,40 @@ color: var(--main-color-dark);
     align-items: center;
     justify-content: flex-end;
 }
+
+/*LOGIN AND ACC CREATION*/
+#login-form {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    background-color: red;
+    padding: 10px;
+    border-radius: 16px;
+    width: 50vw;
+  }
+  
+  #login-form>input {
+    border-radius: 8px;
+    border: none;
+  }
+  
+  #create-account-button {
+    background-color: aqua;
+    color: red;
+    border: none;
+    border-radius: 8px;
+  }
+  
+  #login-modal-bg {
+    position: absolute;
+    height: 100vh;
+    width: 100vw;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 10;
+  }
 }
 
 /*SMALL SCREENS*/
